backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
ccxt==4.1.13
pandas==2.1.3
numpy==1.26.2
websockets==12.0
aiohttp==3.9.1
pytest==7.4.3
alembic==1.12.1
psycopg2-binary==2.9.9

# 업비트 API 연동
pyupbit==0.2.33
cryptography==41.0.5

httpx==0.25.2
starlette==0.27.0

pydantic-settings==2.1.0
